+++ src-bin/Main.hs
@@ -15,7 +15,6 @@ import qualified Clay as C
 import Development.Shake
 import Lucid
 import Main.Utf8
-import Neuron.Version (neuronVersion, olderThan)
 import qualified Neuron.Zettelkasten as Z
 import qualified Neuron.Zettelkasten.Config as Z
 import qualified Neuron.Zettelkasten.Route as Z
@@ -31,11 +30,9 @@ generateSite :: Action ()
 generateSite = do
   Rib.buildStaticFiles ["static/**"]
   config <- Z.getConfig
-  when (olderThan $ Z.minVersion config) $ do
-    error $ "Require neuron mininum version " <> Z.minVersion config <> ", but your neuron version is " <> neuronVersion
   let writeHtmlRoute :: Z.Route s g a -> (s, g, a) -> Action ()
       writeHtmlRoute r = Rib.writeRoute r . Lucid.renderText . renderPage config r
-  void $ Z.generateSite writeHtmlRoute ["*.md"]
+  void $ Z.generateSite config writeHtmlRoute ["*.md"]
 
 renderPage :: Z.Config -> Z.Route s g a -> (s, g, a) -> Html ()
 renderPage config r val@(s, _, _) = html_ [lang_ "en"] $ do