+++ test/Neuron/Zettelkasten/ConfigSpec.hs
@@ -1,0 +1,31 @@
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE QuasiQuotes #-}
+{-# LANGUAGE RecordWildCards #-}
+{-# LANGUAGE ViewPatterns #-}
+{-# LANGUAGE NoImplicitPrelude #-}
+
+module Neuron.Zettelkasten.ConfigSpec
+  ( spec,
+  )
+where
+
+import Neuron.Parser
+import qualified Neuron.Zettelkasten.Config as Z
+import qualified Neuron.Zettelkasten.ID as Z
+import Relude
+import Test.Hspec
+
+spec :: Spec
+spec = do
+  describe "Alias parsing" $ do
+    itParsesAlias "with z-index" "index:z-index"
+    itParsesAlias "with normal id" "foo:2011501"
+    itParsesAlias "longish alias" "tis-a-furphy:2011501"
+
+itParsesAlias :: String -> Text -> SpecWith ()
+itParsesAlias name s =
+  it name $ do
+    fmap renderAlias (parse Z.aliasParser "<hspec>" s) `shouldBe` Right s
+  where
+    renderAlias Z.Alias {..} =
+      Z.zettelIDText aliasZettel <> ":" <> Z.zettelIDText targetZettel